requirements.txt
python-dotenv==1.0.1
telethon==1.38.1
